+++ modules/router/src/services/transfer.ts
@@ -84,7 +84,10 @@ export const attemptTransferWithCollateralization = async (
   // check if recipient is available
   let available = false;
   try {
-    const online = await nodeService.sendIsAliveMessage({ channelAddress: params.channelAddress, skipCheckIn: true });
+    const online = await nodeService.sendIsAliveMessage({
+      channelAddress: params.channelAddress,
+      skipCheckIn: true,
+    });
     available = !online.isError;
   } catch (e) {
     logger.warn(