+++ modules/documentation/docs/changelog.md
@@ -1,6 +1,22 @@
 # Vector Changelog
 
 ## Next Release
+## 0.1.8-beta.2
+
+- \[messaging\] Fix for loop in message retry code
+
+## 0.1.8-beta.1
+
+- \[protocol\] Remove retrying on protocol messages
+- \[protocol\] Hard error for lock release
+
+(Kids -- don't debug on prod)
+
+## 0.1.8-beta.0
+
+- \[protocol\] Retry sending messages
+- \[protocol\] Do not throw error if lock is not released
+- \[messaging\] Add retry logic
 
 ## 0.1.7